+++ store/v2/commitment/tree.go
@@ -6,6 +6,7 @@ import (
 
 	ics23 "github.com/cosmos/ics23/go"
 
+	corestore "cosmossdk.io/core/store"
 	snapshotstypes "cosmossdk.io/store/v2/snapshots/types"
 )
 
@@ -29,17 +30,18 @@ type Tree interface {
 	SetInitialVersion(version uint64) error
 	GetProof(version uint64, key []byte) (*ics23.CommitmentProof, error)
 
-	// Get attempts to retrieve a value from the tree for a given version.
-	//
-	// NOTE: This method only exists to support migration from IAVL v0/v1 to v2.
-	// Once migration is complete, this method should be removed and/or not used.
-	Get(version uint64, key []byte) ([]byte, error)
-
 	Prune(version uint64) error
 	Export(version uint64) (Exporter, error)
 	Import(version uint64) (Importer, error)
 
 	io.Closer
+}
+
+// Reader is the optional interface that is only used to read data from the tree
+// during the migration process.
+type Reader interface {
+	Get(version uint64, key []byte) ([]byte, error)
+	Iterator(version uint64, start, end []byte, ascending bool) (corestore.Iterator, error)
 }
 
 // Exporter is the interface that wraps the basic Export methods.