+++ x/accounts/msg_server.go
@@ -2,6 +2,7 @@ package accounts
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"cosmossdk.io/core/event"
@@ -10,6 +11,8 @@ import (
 )
 
 var _ v1.MsgServer = msgServer{}
+
+var ErrBundlingDisabled = errors.New("accounts: bundling is disabled")
 
 func NewMsgServer(k Keeper) v1.MsgServer {
 	return &msgServer{k}
@@ -85,5 +88,18 @@ func (m msgServer) Execute(ctx context.Context, execute *v1.MsgExecute) (*v1.Msg
 }
 
 func (m msgServer) ExecuteBundle(ctx context.Context, req *v1.MsgExecuteBundle) (*v1.MsgExecuteBundleResponse, error) {
-	panic("impl")
+	if m.k.bundlingDisabled {
+		return nil, ErrBundlingDisabled
+	}
+
+	_, err := m.k.addressCodec.StringToBytes(req.Bundler)
+	if err != nil {
+		return nil, err
+	}
+	responses := make([]*v1.BundledTxResponse, len(req.Txs))
+	for i, bundledTx := range req.Txs {
+		bundleRes := m.k.ExecuteBundledTx(ctx, req.Bundler, bundledTx)
+		responses[i] = bundleRes
+	}
+	return &v1.MsgExecuteBundleResponse{Responses: responses}, nil
 }